requirements.txt
chromadb==0.3.21
Flask==2.3.1
Flask-Cors==3.0.10
langchain==0.0.150
numpy==1.24.3
openai==0.27.4
python-dotenv==1.0.0
tiktoken==0.3.3
